Two changes made for better support for frozen columns with column reordering enabled.
slick.draggablegrouping.js 
Only one header's ids were being accounted for after a drag and drop reorder when columns were frozen. the result is that the left (frozen) columns were fine, but the right (unfrozen) columns were 1) not sorted, and b) turned off.

slick.grid.js
Similar to the draggablegrouping issue, the grid was not honoring the preheader state for the unfrozen columns because only hte preheaderPanelScroller was  being updated setPreHeaderPanelVisibility. For me this showed itself because I had teh preheader row toggled off and froze some columns and the headers disappeared for my unfrozen columns due to the preheader row being visible but empty and the headers being under the first row of data.