To Fix the issue specific to drag selection in non-frozen part of grid if there are multiple grids.
Issue Description: if we have multiple grids in same page/in the DOM, the above code looks for  '.grid-canvas-left' width which indeed gets the result of the grid where the event triggered and the other grids present in the DOM. If the action is done on second grid(which has frozen column) instead of first one above code is getting the width of first grid instead of second one. So made changes to get the width only for the event triggered grid.