fix: allow for multiple css class names (#795)
The previous behaviour with jQuery was that anywhere a class
name was passed, jQuery allowed a space separated list of
class names.
With "classList.add" this behaviour has to be explicitly added.

I believe all cases where ".split()" is used are already guarded
against the classname not being falsy.
There were a couple of places where ".split" was used, but not
the spread operator (...). This would result in a comma separated
list of classnames being added. The assumption is that was a bug
so I have fixed those.

I haven't changed any code in "dist", I assume that is handled
as part of the release process in some way.