Fix typing - GroupingGetterFunction must be allowed to return arbitrary value
Grouping logic groups items based on equality of returned grouping values;
probably does not need to be string. For example, we are using it to group based
on Date value (by e.g. rounding them to nearest whole hour). If string is
required by the grouping function, additional conversion of Date to string is
needed, which we would rather avoid for perf reasons.